+++ app/models/concerns/noticed/deliverable.rb
@@ -38,6 +38,13 @@ module Noticed
       def deliver_by(name, options = {})
         raise NameError, "#{name} has already been used for this Notifier." if delivery_methods.has_key?(name)
 
+        if name == :database
+          Noticed.deprecator.warn <<-WARNING.squish
+            The :database delivery method has been deprecated and does nothing. Notifiers automatically save to the database now.
+          WARNING
+          return
+        end
+
         config = ActiveSupport::OrderedOptions.new.merge(options)
         yield config if block_given?
         delivery_methods[name] = DeliverBy.new(name, config)
@@ -48,6 +55,20 @@ module Noticed
       end
       alias_method :required_param, :required_params
 
+      def params(*names)
+        Noticed.deprecator.warn <<-WARNING.squish
+          `params` is deprecated and has been renamed to `required_params`
+        WARNING
+        required_params(*names)
+      end
+
+      def param(*names)
+        Noticed.deprecator.warn <<-WARNING.squish
+          `param :name` is deprecated and has been renamed to `required_param :name`
+        WARNING
+        required_params(*names)
+      end
+
       def with(params)
         record = params.delete(:record)
         new(params: params, record: record)
@@ -56,6 +77,7 @@ module Noticed
       def deliver(recipients = nil, options = {})
         new.deliver(recipients, options)
       end
+      alias_method :deliver_later, :deliver
     end
 
     # CommentNotifier.deliver(User.all)
@@ -90,11 +112,12 @@ module Noticed
 
       self
     end
+    alias_method :deliver_later, :deliver
 
     def recipient_attributes_for(recipient)
       {
         type: "#{self.class.name}::Notification",
-        recipient_type: recipient.class.name,
+        recipient_type: recipient.class.base_class.name,
         recipient_id: recipient.id
       }
     end